Move SearchDatasetTutorial component
